Se aplico exception a EquipoServiceImpl

+++ src/main/java/com/torneos/futbol/service/EquipoServiceImpl.java
@@ -1,5 +1,6 @@
 package com.torneos.futbol.service;
 
+import com.torneos.futbol.exception.BadRequestException;
 import com.torneos.futbol.model.dto.EquipoDto;
 import com.torneos.futbol.model.entity.Equipo;
 import com.torneos.futbol.model.entity.Torneo;
@@ -9,6 +10,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
@@ -18,28 +20,43 @@ public class EquipoServiceImpl  implements EquipoService {
     private final EquipoRepository equipoRepository;
     private  final TorneoRepository torneoRepository;
     @Override
-    public Equipo save(EquipoDto equipoDto) {
+    public Equipo save(EquipoDto equipoDto) throws BadRequestException {
         Equipo equipo = new Equipo();
         equipo.setNombre(equipoDto.getNombre());
-
-        Integer torneoId = equipoDto.getTorneoId();
-
-        if (torneoId != null && torneoRepository.existsById(torneoId)) {
-            Torneo torneo = torneoRepository.findById(torneoId).orElse(null);
-            equipo.setTorneo(torneo);
+        if (equipoDto.getTorneoId() != null) {
+            Optional<Torneo> torneoOptional = torneoRepository.findById(equipoDto.getTorneoId());
+            if (torneoOptional.isPresent()){
+                equipo.setTorneo(torneoOptional.get());
+            }
+            else {
+                throw new BadRequestException("El Torneo con Id: "+equipoDto.getTorneoId()+ " no existe");
+            }
         }
 
         return equipoRepository.save(equipo);
     }
 
     @Override
-    public void delete(Equipo equipo) {
-        equipoRepository.delete(equipo);
+    public void delete(Integer id) throws BadRequestException {
+        Optional<Equipo> equipoOptional = equipoRepository.findById(id);
+        if (equipoOptional.isPresent()) {
+            Equipo equipo = equipoOptional.get();
+            equipoRepository.delete(equipo);
+        }
+        else {
+            throw new BadRequestException("No se encontro el Equipo con el Id: " +id);
+        }
     }
 
     @Override
-    public Equipo findById(Integer id) {
-        return equipoRepository.findById(id).orElse(null);
+    public Equipo findById(Integer id) throws BadRequestException {
+        Optional<Equipo> equipoOptional = equipoRepository.findById(id);
+        if (equipoOptional.isPresent()){
+            return equipoOptional.get();
+        }
+        else {
+            throw new BadRequestException("No se encuentra ningun Equipo con Id: " +id);
+        }
     }
 
     @Override
@@ -48,16 +65,29 @@ public class EquipoServiceImpl  implements EquipoService {
     }
 
     @Override
-    public Equipo update(EquipoDto equipoDto,Integer id) {
+    public Equipo update(EquipoDto equipoDto,Integer id) throws BadRequestException {
 
-        Equipo equipo = new Equipo();
-        equipo.setNombre(equipoDto.getNombre());
-
-        if(equipoDto.getTorneoId() !=null && torneoRepository.existsById(equipoDto.getTorneoId())){
-            Torneo torneo = torneoRepository.findById(equipoDto.getTorneoId()).orElse(null);
-            equipo.setTorneo(torneo);
+        if(equipoDto != null) {
+            Optional<Equipo> equipoOptional = equipoRepository.findById(id);
+            if(equipoOptional.isPresent()) {
+                Equipo equipo = equipoOptional.get();
+                equipo.setNombre(equipoDto.getNombre());
+                Optional<Torneo> torneoOptional = torneoRepository.findById(equipoDto.getTorneoId());
+                if (torneoOptional.isPresent())
+                {
+                    equipo.setTorneo(torneoOptional.get());
+                }
+                else {
+                    throw new BadRequestException("El Torneo con Id: " +equipoDto.getTorneoId() + " no existe");
+                }
+                return equipoRepository.save(equipo);
+            }
+            else {
+                throw new BadRequestException("El Equipo con Id: " +id + " no existe");
+            }
         }
-        equipo.setId(id);
-        return  equipoRepository.save(equipo);
+        else {
+            throw new BadRequestException("El Equipo no puede ser null");
+        }
     }
 }
